+++ src/tssl/echo.h
@@ -10,8 +10,8 @@ namespace ts {
 	class Engine;
 	
 	namespace sl {
-		Entry* echo(Engine* engine, size_t argc, Entry* args);
-		Entry* error(Engine* engine, size_t argc, Entry* args);
+		Entry* echo(Engine* engine, unsigned int argc, Entry* args);
+		Entry* error(Engine* engine, unsigned int argc, Entry* args);
 
 		extern string mockStdout;
 		extern bool useMockStdout;
