+++ src/tssl/eval.cc
@@ -4,7 +4,7 @@
 
 namespace ts {
 	namespace sl {
-		Entry* eval(Engine* engine, size_t argc, Entry* args) {
+		Entry* eval(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				string shell(args[0].stringData);
 				engine->execShell(shell, true);
