+++ src/tssl/exec.cc
@@ -4,7 +4,7 @@
 
 namespace ts {
 	namespace sl {
-		Entry* exec(Engine* engine, size_t argc, Entry* args) {
+		Entry* exec(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				string fileName(args[0].stringData);
 				engine->execFile(fileName, true);
