+++ src/tssl/exec.h
@@ -8,6 +8,6 @@ namespace ts {
 	class Engine;
 	
 	namespace sl {
-		Entry* exec(Engine* engine, size_t argc, Entry* args);
+		Entry* exec(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
