+++ src/tssl/fileObject.cc
@@ -58,7 +58,7 @@ namespace ts {
 			wrapper->data = new FileObject();
 		}
 
-		Entry* FileObject__openForRead(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__openForRead(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 2 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				((FileObject*)args[0].objectData->objectWrapper->data)->open(args[1].stringData, READ);
 			}
@@ -66,7 +66,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__openForWrite(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__openForWrite(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 2 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				((FileObject*)args[0].objectData->objectWrapper->data)->open(args[1].stringData, WRITE);
 			}
@@ -74,7 +74,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__openForAppend(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__openForAppend(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 2 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				((FileObject*)args[0].objectData->objectWrapper->data)->open(args[1].stringData, APPEND);
 			}
@@ -82,7 +82,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__close(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__close(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				((FileObject*)args[0].objectData->objectWrapper->data)->close();
 			}
@@ -90,7 +90,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__readLine(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__readLine(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				return new Entry(((FileObject*)args[0].objectData->objectWrapper->data)->readLine());
 			}
@@ -98,7 +98,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__writeLine(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__writeLine(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 2 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				((FileObject*)args[0].objectData->objectWrapper->data)->writeLine(args[1].stringData);
 			}
@@ -106,7 +106,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* FileObject__isEOF(Engine* engine, size_t argc, Entry* args) {
+		Entry* FileObject__isEOF(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1 && args[0].objectData->objectWrapper->object->typeMethodTree->name == "FileObject") {
 				return new Entry((double)((FileObject*)args[0].objectData->objectWrapper->data)->isEOF());
 			}
@@ -114,7 +114,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* fileBase(Engine* engine, size_t argc, Entry* args) {
+		Entry* fileBase(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				const char* path = args[0].stringData;
 				const char* slashLocation = strrchr(path, '/');
@@ -141,7 +141,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* fileExt(Engine* engine, size_t argc, Entry* args) {
+		Entry* fileExt(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				const char* path = args[0].stringData;
 				const char* dotLocation = strrchr(path, '.');
@@ -155,7 +155,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* fileName(Engine* engine, size_t argc, Entry* args) {
+		Entry* fileName(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				const char* path = args[0].stringData;
 				const char* slashLocation = strrchr(path, '/');
@@ -172,7 +172,7 @@ namespace ts {
 			return nullptr;
 		}
 
-		Entry* filePath(Engine* engine, size_t argc, Entry* args) {
+		Entry* filePath(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				const char* path = args[0].stringData;
 				const char* slashLocation = strrchr(path, '/');
