+++ src/tssl/fileObject.h
@@ -33,17 +33,17 @@ namespace ts {
 		};
 
 		void FileObject__constructor(ObjectWrapper* wrapper);
-		Entry* FileObject__openForRead(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__openForWrite(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__openForAppend(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__close(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__readLine(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__writeLine(Engine* engine, size_t argc, Entry* args);
-		Entry* FileObject__isEOF(Engine* engine, size_t argc, Entry* args);
+		Entry* FileObject__openForRead(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__openForWrite(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__openForAppend(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__close(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__readLine(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__writeLine(Engine* engine, unsigned int argc, Entry* args);
+		Entry* FileObject__isEOF(Engine* engine, unsigned int argc, Entry* args);
 
-		Entry* fileBase(Engine* engine, size_t argc, Entry* args);
-		Entry* fileExt(Engine* engine, size_t argc, Entry* args);
-		Entry* fileName(Engine* engine, size_t argc, Entry* args);
-		Entry* filePath(Engine* engine, size_t argc, Entry* args);
+		Entry* fileBase(Engine* engine, unsigned int argc, Entry* args);
+		Entry* fileExt(Engine* engine, unsigned int argc, Entry* args);
+		Entry* fileName(Engine* engine, unsigned int argc, Entry* args);
+		Entry* filePath(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
