+++ src/tssl/getField.h
@@ -11,12 +11,12 @@ namespace ts {
 	class Engine;
 
 	namespace sl {
-		Entry* firstField(Engine* engine, size_t argc, Entry* args);
-		Entry* restFields(Engine* engine, size_t argc, Entry* args);
-		Entry* getField(Engine* engine, size_t argc, Entry* args);
-		Entry* getFields(Engine* engine, size_t argc, Entry* args);
-		Entry* getFieldCount(Engine* engine, size_t argc, Entry* args);
-		Entry* removeField(Engine* engine, size_t argc, Entry* args);
-		Entry* setField(Engine* engine, size_t argc, Entry* args);
+		Entry* firstField(Engine* engine, unsigned int argc, Entry* args);
+		Entry* restFields(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getField(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getFields(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getFieldCount(Engine* engine, unsigned int argc, Entry* args);
+		Entry* removeField(Engine* engine, unsigned int argc, Entry* args);
+		Entry* setField(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
