+++ src/tssl/getRecord.h
@@ -11,12 +11,12 @@ namespace ts {
 	class Engine;
 
 	namespace sl {
-		Entry* firstRecord(Engine* engine, size_t argc, Entry* args);
-		Entry* restRecords(Engine* engine, size_t argc, Entry* args);
-		Entry* getRecord(Engine* engine, size_t argc, Entry* args);
-		Entry* getRecords(Engine* engine, size_t argc, Entry* args);
-		Entry* getRecordCount(Engine* engine, size_t argc, Entry* args);
-		Entry* removeRecord(Engine* engine, size_t argc, Entry* args);
-		Entry* setRecord(Engine* engine, size_t argc, Entry* args);
+		Entry* firstRecord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* restRecords(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getRecord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getRecords(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getRecordCount(Engine* engine, unsigned int argc, Entry* args);
+		Entry* removeRecord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* setRecord(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
