+++ src/tssl/getWord.h
@@ -11,12 +11,12 @@ namespace ts {
 	class Engine;
 
 	namespace sl {
-		Entry* firstWord(Engine* engine, size_t argc, Entry* args);
-		Entry* restWords(Engine* engine, size_t argc, Entry* args);
-		Entry* getWord(Engine* engine, size_t argc, Entry* args);
-		Entry* getWords(Engine* engine, size_t argc, Entry* args);
-		Entry* getWordCount(Engine* engine, size_t argc, Entry* args);
-		Entry* removeWord(Engine* engine, size_t argc, Entry* args);
-		Entry* setWord(Engine* engine, size_t argc, Entry* args);
+		Entry* firstWord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* restWords(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getWord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getWords(Engine* engine, unsigned int argc, Entry* args);
+		Entry* getWordCount(Engine* engine, unsigned int argc, Entry* args);
+		Entry* removeWord(Engine* engine, unsigned int argc, Entry* args);
+		Entry* setWord(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
