+++ src/tssl/isObject.cc
@@ -6,7 +6,7 @@ using namespace std;
 
 namespace ts {
 	namespace sl {
-		Entry* isObject(Engine* engine, size_t argc, Entry* args) {
+		Entry* isObject(Engine* engine, unsigned int argc, Entry* args) {
 			if(argc == 1) {
 				if(args[0].objectData->objectWrapper != nullptr) {
 					return new Entry(1.0);
