+++ src/tssl/schedule.h
@@ -6,8 +6,8 @@ namespace ts {
 	class Engine;
 
 	namespace sl {
-		Entry* schedule(Engine* engine, size_t argc, Entry* args);
+		Entry* schedule(Engine* engine, unsigned int argc, Entry* args);
 
-		Entry* SimObject__schedule(Engine* engine, size_t argc, Entry* args);
+		Entry* SimObject__schedule(Engine* engine, unsigned int argc, Entry* args);
 	}
 }
